Rm wait util and make abstract methods abstract
